alloy_network
Module any

Types for handling unknown network types.

Structs

AnyHeader

Block header representation with certain fields made optional to account for possible differences in network implementations.

AnyNetwork

Types for a catch-all network.

AnyReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

AnyRpcBlock

A wrapper for AnyRpcBlock that allows for handling unknown block types.

AnyRpcTransaction

A wrapper for AnyRpcTransaction that allows for handling unknown transaction types.

AnyTxType

Transaction type for a catch-all network.

UnknownTxEnvelope

A transaction envelope from an unknown network.

UnknownTypedTransaction

A typed transaction of an unknown Network

Enums

AnyTxEnvelope

Transaction envelope for a catch-all network.

AnyTypedTransaction

Unsigned transaction type for a catch-all network.

Type Aliases

AnyRpcHeader

A catch-all header type for handling headers on multiple networks.

AnyTransactionReceipt

Alias for a catch-all receipt type.