alloy_network::eip2718

Trait Typed2718

pub trait Typed2718 {
    // Required method
    fn ty(&self) -> u8;

    // Provided methods
    fn is_type(&self, ty: u8) -> bool { ... }
    fn is_legacy(&self) -> bool { ... }
    fn is_eip2930(&self) -> bool { ... }
    fn is_eip1559(&self) -> bool { ... }
    fn is_eip4844(&self) -> bool { ... }
    fn is_eip7702(&self) -> bool { ... }
}

A trait that helps to determine the type of the transaction.

Required Methods

fn ty(&self) -> u8

Returns the EIP-2718 type flag.

Provided Methods

fn is_type(&self, ty: u8) -> bool

Returns true if the type matches the given type.

fn is_legacy(&self) -> bool

Returns true if the type is a legacy transaction.

fn is_eip2930(&self) -> bool

Returns true if the type is an EIP-2930 transaction.

fn is_eip1559(&self) -> bool

Returns true if the type is an EIP-1559 transaction.

fn is_eip4844(&self) -> bool

Returns true if the type is an EIP-4844 transaction.

fn is_eip7702(&self) -> bool

Returns true if the type is an EIP-7702 transaction.