alloy_network

Trait FullSigner

pub trait FullSigner<S>: Signer<S> + TxSigner<S> { }

A unifying trait for asynchronous Ethereum signers that combine the functionalities of both Signer and TxSigner.

This trait enables dynamic dispatch (e.g., using Box<dyn FullSigner>) for types that combine both asynchronous Ethereum signing and transaction signing functionalities.