Trait alloy_network::Network

pub trait Network: Clone + Copy + Sized + Send + Sync + 'static {
    type TxEnvelope: Eip2718Envelope;
    type UnsignedTx;
    type ReceiptEnvelope: Eip2718Envelope;
    type Header;
    type TransactionRequest: RpcObject + TransactionBuilder<Self> + Debug;
    type TransactionResponse: RpcObject;
    type ReceiptResponse: RpcObject + ReceiptResponse;
    type HeaderResponse: RpcObject;
}

Captures type info for network-specific RPC requests/responses.

Networks are only containers for types, so it is recommended to use ZSTs for their definition.

Required Associated Types

type TxEnvelope: Eip2718Envelope

The network transaction envelope type.

type UnsignedTx

An enum over the various transaction types.

type ReceiptEnvelope: Eip2718Envelope

The network receipt envelope type.

type Header

The network header type.

type TransactionRequest: RpcObject + TransactionBuilder<Self> + Debug

The JSON body of a transaction request.

type TransactionResponse: RpcObject

The JSON body of a transaction response.

type ReceiptResponse: RpcObject + ReceiptResponse

The JSON body of a transaction receipt.

type HeaderResponse: RpcObject

The JSON body of a header response, as flattened into BlockResponse.

Object Safety

This trait is not object safe.