Trait alloy_network::Receipt

pub trait Receipt {
    // Required methods
    fn success(&self) -> bool;
    fn bloom(&self) -> Bloom;
    fn cumulative_gas_used(&self) -> u64;
    fn logs(&self) -> &[Log];

    // Provided method
    fn bloom_cheap(&self) -> Option<Bloom> { ... }
}

Receipt is the result of a transaction execution.

Required Methods

fn success(&self) -> bool

Returns true if the transaction was successful.

fn bloom(&self) -> Bloom

Returns the bloom filter for the logs in the receipt. This operation may be expensive.

fn cumulative_gas_used(&self) -> u64

Returns the cumulative gas used in the block after this transaction was executed.

fn logs(&self) -> &[Log]

Returns the logs emitted by this transaction.

Provided Methods

fn bloom_cheap(&self) -> Option<Bloom>

Returns the bloom filter for the logs in the receipt, if it is cheap to compute.