Trait alloy_network::ReceiptResponse

pub trait ReceiptResponse {
    // Required methods
    fn contract_address(&self) -> Option<Address>;
    fn status(&self) -> bool;
}

A receipt response.

This is distinct from TxReceipt, since this is for JSON-RPC receipts.

Required Methods

fn contract_address(&self) -> Option<Address>

Address of the created contract, or None if the transaction was not a deployment.

fn status(&self) -> bool

Status of the transaction.

Note

Caution must be taken when using this method for deep-historical receipts, as it may not accurately reflect the status of the transaction. The transaction status is not knowable from the receipt for transactions before EIP-658.

This can be handled using TxReceipt::status_or_post_state.