Trait alloy_network::Transaction

pub trait Transaction: Any + Encodable + Send + Sync + 'static {
    type Signature;

    // Required methods
    fn into_signed(self, signature: Signature) -> Signed<Self, Self::Signature>
       where Self: Sized;
    fn encode_signed(&self, signature: &Signature, out: &mut dyn BufMut);
    fn decode_signed(buf: &mut &[u8]) -> Result<Signed<Self>>
       where Self: Sized;
    fn signature_hash(&self) -> B256;
    fn input(&self) -> &[u8] ;
    fn input_mut(&mut self) -> &mut Bytes;
    fn set_input(&mut self, data: Bytes);
    fn to(&self) -> TxKind;
    fn set_to(&mut self, to: TxKind);
    fn value(&self) -> U256;
    fn set_value(&mut self, value: U256);
    fn chain_id(&self) -> Option<ChainId>;
    fn set_chain_id(&mut self, chain_id: ChainId);
    fn nonce(&self) -> u64;
    fn set_nonce(&mut self, nonce: u64);
    fn gas_limit(&self) -> u64;
    fn set_gas_limit(&mut self, limit: u64);
    fn gas_price(&self) -> Option<U256>;
    fn set_gas_price(&mut self, price: U256);

}

Represents a minimal EVM transaction.

Required Associated Types

type Signature

The signature type for this transaction.

This is usually alloy_primitives::Signature, however, it may be different for future EIP-2718 transaction types, or in other networks. For example, in Optimism, the deposit transaction signature is the unit type ().

Required Methods

fn into_signed(self, signature: Signature) -> Signed<Self, Self::Signature>
where Self: Sized,

Convert to a signed transaction by adding a signature and computing the hash.

fn encode_signed(&self, signature: &Signature, out: &mut dyn BufMut)

Encode with a signature. This encoding is usually RLP, but may be different for future EIP-2718 transaction types.

fn decode_signed(buf: &mut &[u8]) -> Result<Signed<Self>>
where Self: Sized,

Decode a signed transaction. This decoding is usually RLP, but may be different for future EIP-2718 transaction types.

This MUST be the inverse of Transaction::encode_signed.

fn signature_hash(&self) -> B256

Calculate the signing hash for the transaction.

fn input(&self) -> &[u8]

Get data.

fn input_mut(&mut self) -> &mut Bytes

Get data.

fn set_input(&mut self, data: Bytes)

Set data.

fn to(&self) -> TxKind

Get to.

fn set_to(&mut self, to: TxKind)

Set to.

fn value(&self) -> U256

Get value.

fn set_value(&mut self, value: U256)

Set value.

fn chain_id(&self) -> Option<ChainId>

Get chain_id.

fn set_chain_id(&mut self, chain_id: ChainId)

Set chain_id.

fn nonce(&self) -> u64

Get nonce.

fn set_nonce(&mut self, nonce: u64)

Set nonce.

fn gas_limit(&self) -> u64

Get gas_limit.

fn set_gas_limit(&mut self, limit: u64)

Set gas_limit.

fn gas_price(&self) -> Option<U256>

Get gas_price.

fn set_gas_price(&mut self, price: U256)

Set gas_price.