alloy_network

Trait TransactionBuilder4844

pub trait TransactionBuilder4844:
    Sized
    + Default
    + Send
    + Sync
    + 'static {
    // Required methods
    fn max_fee_per_blob_gas(&self) -> Option<u128>;
    fn set_max_fee_per_blob_gas(&mut self, max_fee_per_blob_gas: u128);
    fn blob_sidecar(&self) -> Option<&BlobTransactionSidecar>;
    fn set_blob_sidecar(&mut self, sidecar: BlobTransactionSidecar);

    // Provided methods
    fn with_max_fee_per_blob_gas(self, max_fee_per_blob_gas: u128) -> Self { ... }
    fn with_blob_sidecar(self, sidecar: BlobTransactionSidecar) -> Self { ... }
}

Transaction builder type supporting EIP-4844 transaction fields.

Required Methods

fn max_fee_per_blob_gas(&self) -> Option<u128>

Get the max fee per blob gas for the transaction.

fn set_max_fee_per_blob_gas(&mut self, max_fee_per_blob_gas: u128)

Set the max fee per blob gas for the transaction.

fn blob_sidecar(&self) -> Option<&BlobTransactionSidecar>

Gets the EIP-4844 blob sidecar of the transaction.

fn set_blob_sidecar(&mut self, sidecar: BlobTransactionSidecar)

Sets the EIP-4844 blob sidecar of the transaction.

Note: This will also set the versioned blob hashes accordingly: BlobTransactionSidecar::versioned_hashes

Provided Methods

fn with_max_fee_per_blob_gas(self, max_fee_per_blob_gas: u128) -> Self

Builder-pattern method for setting max fee per blob gas .

fn with_blob_sidecar(self, sidecar: BlobTransactionSidecar) -> Self

Builder-pattern method for setting the EIP-4844 blob sidecar of the transaction.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.