Crate alloy_network

alloy-networks

Ethereum blockchain RPC behavior abstraction.

This crate contains a simple abstraction of the RPC behavior of an Ethereum-like blockchain. It is intended to be used by the Alloy client to provide a consistent interface to the rest of the library, regardless of changes the underlying blockchain makes to the RPC interface.

Core Model

This crate handles abstracting RPC types. It does not handle the actual networking. The core model is as follows:

Transaction - A trait that defines an abstract interface for EVM-like transactions.

Network - A trait that defines the RPC types for a given blockchain. Providers are parameterized by a Network type, and use the associated types to define the input and output types of the RPC methods.

TransactionBuilder - A trait for constructing and validating network-specific transaction requests. Used to build typed transactions for signing and submission. See TransactionBuilder.

NetworkWallet - A trait for wallets that can sign transactions for a given network. Used to abstract over different signing backends. See NetworkWallet.

BlockResponse, TransactionResponse, ReceiptResponse, HeaderResponse - Traits (from alloy-network-primitives) that define the structure of block, transaction, receipt, and header types used in RPC responses. These are associated types in the Network trait and are implemented by network-specific types. See alloy-network-primitives.

Usage

This crate is not intended to be used directly. It is used by the alloy-provider library and reth to modify the input and output types of the RPC methods.

This crate will primarily be used by blockchain maintainers to add bespoke RPC types to the Alloy provider. This is done by implementing the Network trait, and then parameterizing the Provider type with the new network type.

For example, to add a new network called Foo:

// Foo must be a ZST. It is a compile error to use a non-ZST type.
struct Foo;

impl Network for Foo {
    type Transaction = FooTransaction;
    type Block = FooBlock;
    type Header = FooHeader;
    type Receipt = FooReceipt;

    // etc.
}

The user may then instantiate a Provider<Foo> and use it as normal. This allows the user to use the same API for all networks, regardless of the underlying RPC types.

Note: If you need to also add custom methods to your network, you should make an extension trait for Provider<N> as follows:

#[async_trait]
trait FooProviderExt: Provider<Foo> {
    async fn custom_foo_method(&self) -> RpcResult<Something, TransportError>;

    async fn another_custom_method(&self) -> RpcResult<Something, TransportError>;
}

Re-exports

pub use any::AnyNetwork;

pub use any::AnyRpcBlock;

pub use any::AnyRpcTransaction;

pub use any::AnyTxEnvelope;

pub use any::AnyTxType;

pub use any::AnyTypedTransaction;

pub use any::UnknownTxEnvelope;

pub use any::UnknownTypedTransaction;

pub use alloy_network_primitives as primitives;

Modules

any

Types for handling unknown network types.

eip2718

EIP-2718 traits.

Macros

impl_into_wallet

Utility to implement IntoWallet for signer over the specified network.

Structs

AnyHeader

Block header representation with certain fields made optional to account for possible differences in network implementations.

AnyReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

Ethereum

Types for a mainnet-like Ethereum network.

EthereumWallet

A wallet capable of signing any transaction for the Ethereum network.

UnbuiltTransactionError

An unbuilt transaction, along with some error.

Enums

TransactionBuilderError

Error type for transaction builders.

Traits

BlockResponse

Block JSON-RPC response.

FullSigner

A unifying trait for asynchronous Ethereum signers that combine the functionalities of both Signer and TxSigner.

FullSignerSync

A unifying trait for synchronous Ethereum signers that implement both SignerSync and TxSignerSync.

IntoWallet

A trait for converting a signer into a NetworkWallet.

Network

Captures type info for network-specific RPC requests/responses.

NetworkWallet

A wallet capable of signing any transaction for the given network.

ReceiptResponse

Receipt JSON-RPC response.

TransactionBuilder

A Transaction builder for a network.

TransactionBuilder4844

Transaction builder type supporting EIP-4844 transaction fields.

TransactionBuilder7702

Transaction builder type supporting EIP-7702 transaction fields.

TransactionResponse

Transaction JSON-RPC response. Aggregates transaction data with its block and signer context.

TxSigner

Asynchronous transaction signer, capable of signing any SignableTransaction for the given Signature type.

TxSignerSync

Synchronous transaction signer, capable of signing any SignableTransaction for the given Signature type.

Type Aliases

AnyRpcHeader

A catch-all header type for handling headers on multiple networks.

AnyTransactionReceipt

Alias for a catch-all receipt type.

BuildResult

Result type for transaction builders